Create lgsm-monitor.sh
Log monitor (For POSTGAME) script for an Arma Reforger server running how it's installed via the install instructions on the LGSM website. Script follows the armarserver-console.log file created by the LGSM script. Could be run via tmux to put it in the background, or as a systemd user service depending on your needs.