Update armar-sc.sh
Modified how mods are added to the server json file. Removed the name and version updates to json, and only kept the modId since it's all that is needed. But the script will still display modId/name/version info while in the Manage Mods section so Admins can associate a ModId to a specific mod.

Also changed script version number.