Update server-install-howto.md
Edited the example server JSON. supportedPlatforms is not needed since the server will automatically enable PSN cross play if crossPlatform is set to true, as long as Mods are not enabled. If ModIds are put in the server JSON file, then the server will automatically disable PSN support, but keep PC and XBox.